teacherBa_front/requirements.txt
# LangChain 관련
langchain>=0.1.20
langchain-core>=0.1.35
langchain-community>=0.0.35
langgraph>=0.0.36

# Azure OpenAI (Azure Chat, Embedding 등)
openai>=1.30.0
langchain-openai>=0.1.6

# 벡터 DB
chromadb>=0.4.24

# 문서 로더 (PDF 등)
pypdf>=4.0.1
langchainhub>=0.1.15

# 웹 프레임워크 (API 및 프론트와 연결)
flask>=3.0.0

# 프론트엔드 (선택 사항, 챗봇 시각화용)
streamlit>=1.34.0

# 환경변수 관리
python-dotenv>=1.0.1

# 기타 유틸
tqdm>=4.66.0
numpy>=1.26.4
